+++ client/package.json
@@ -32,9 +32,13 @@
     "@shopify/react-testing": "^5.1.0",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
+    "@typescript-eslint/eslint-plugin": "^6.6.0",
+    "@typescript-eslint/parser": "^6.6.0",
     "@vitejs/plugin-react": "3.1.0",
     "autoprefixer": "^10.4.14",
     "eslint": "^8.35.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-react": "^7.33.2",
     "history": "^5.3.0",
     "jsdom": "^21.1.0",
     "postcss": "^8.4.21",
