+++ client/src/assets/home-trophy.png
