+++ client/src/assets/index.ts
