+++ client/src/hooks/useAuthenticatedFetch.ts
