+++ client/src/lib/general.ts
