+++ client/src/pages/index.tsx
