+++ client/vite.config.ts
@@ -58,6 +58,9 @@ export default defineConfig({
   },
   resolve: {
     preserveSymlinks: true,
+    alias: {
+      "@": resolve(__dirname, "./src"),
+    },
   },
   server: {
     host: "localhost",
